# openapi.yaml
openapi: 3.1.0
info:
  title: GhostTeam API
  version: 0.1.0
  description: GhostTeam API by GodsIMiJ AI Solutions Inc., architected by James D. Ingersoll.
  license:
    name: MIT
servers:
  - url: http://localhost:3000
security:
  - GhostTeamApiKey: []
tags:
  - name: Agents
  - name: Tasks
  - name: Messages
  - name: Logs
  - name: GhostOS
paths:
  /agents:
    get:
      tags: [Agents]
      summary: List agents
      responses:
        "200":
          description: A list of agents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentListResponse"
  /agents/join:
    post:
      tags: [Agents]
      summary: Join an agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinAgentRequest"
      responses:
        "201":
          description: Joined agent
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/AgentResponseWrapper"
                  - $ref: "#/components/schemas/IdNoteResponse"
  /agents/leave:
    post:
      tags: [Agents]
      summary: Leave an agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeaveAgentRequest"
      responses:
        "200":
          description: Agent removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveAgentResponse"
  /agents/{id}:
    get:
      tags: [Agents]
      summary: Get agent details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Agent details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentResponseWrapper"
        "404":
          description: Agent not found
  /tasks:
    get:
      tags: [Tasks]
      summary: List tasks
      responses:
        "200":
          description: A list of tasks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskListResponse"
  /tasks/create:
    post:
      tags: [Tasks]
      summary: Create a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskRequest"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/TaskDetailsResponseWrapper"
                  - $ref: "#/components/schemas/IdNoteResponse64"
  /tasks/ack:
    post:
      tags: [Tasks]
      summary: Acknowledge a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AckTaskRequest"
      responses:
        "200":
          description: Task acknowledged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskStatusResponse"
  /tasks/complete:
    post:
      tags: [Tasks]
      summary: Complete a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteTaskRequest"
      responses:
        "200":
          description: Task completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskStatusResponse"
  /tasks/requeue:
    post:
      tags: [Tasks]
      summary: Requeue a task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequeueTaskRequest"
      responses:
        "200":
          description: Task requeued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskStatusResponse"
  /tasks/{id}:
    get:
      tags: [Tasks]
      summary: Get task details and history
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Task details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskDetailsResponseWrapper"
        "404":
          description: Task not found
  /messages/{agent}:
    get:
      tags: [Messages]
      summary: Get unread messages for an agent
      parameters:
        - name: agent
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Unread messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageListResponse"
  /messages/send:
    post:
      tags: [Messages]
      summary: Send a message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageRequest"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageSendResponse"
  /messages/mark-read:
    post:
      tags: [Messages]
      summary: Mark a message as read
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkReadRequest"
      responses:
        "200":
          description: Message marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageReadResponse"
  /logs/{agent}/stream:
    get:
      tags: [Logs]
      summary: Stream agent logs over WebSocket
      parameters:
        - name: agent
          in: path
          required: true
          schema:
            type: string
      responses:
        "101":
          description: WebSocket upgrade for streaming log lines
  /ghostos/infer:
    post:
      tags: [GhostOS]
      summary: Forward a prompt to GhostOS inference
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GhostOsInferRequest"
      responses:
        "200":
          description: GhostOS inference output
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GhostOsInferResponse"
components:
  securitySchemes:
    GhostTeamApiKey:
      type: apiKey
      in: header
      name: X-GhostTeam-Key
  schemas:
    Agent:
      type: object
      properties:
        id:
          type: string
        role:
          type: string
        backend:
          type: string
        joined_at:
          type:
            - string
            - "null"
      required: [id, role, backend]
    AgentResponseWrapper:
      type: object
      properties:
        ok:
          type: boolean
          const: true
        data:
          $ref: "#/components/schemas/Agent"
      required: [ok, data]
    AgentListResponse:
      type: object
      properties:
        ok:
          type: boolean
          const: true
        data:
          type: array
          items:
            $ref: "#/components/schemas/Agent"
      required: [ok, data]
    JoinAgentRequest:
      type: object
      properties:
        id:
          type: string
        role:
          type: string
        backend:
          type: string
      required: [id, role, backend]
    LeaveAgentRequest:
      type: object
      properties:
        id:
          type: string
      required: [id]
    LeaveAgentResponse:
      type: object
      properties:
        ok:
          type: boolean
        id:
          type: string
      required: [ok, id]
    IdNoteResponse:
      type: object
      properties:
        ok:
          type: boolean
        id:
          type: string
        note:
          type:
            - string
            - "null"
        warning:
          type:
            - string
            - "null"
      required: [ok, id]
    IdNoteResponse64:
      type: object
      properties:
        ok:
          type: boolean
        id:
          type: integer
          format: int64
        note:
          type:
            - string
            - "null"
        warning:
          type:
            - string
            - "null"
      required: [ok, id]
    Task:
      type: object
      properties:
        id:
          type: integer
          format: int64
        creator:
          type: string
        assignee:
          type:
            - string
            - "null"
        description:
          type: string
        status:
          type: string
        result:
          type:
            - string
            - "null"
        created_at:
          type:
            - string
            - "null"
        updated_at:
          type:
            - string
            - "null"
      required: [id, creator, description, status]
    TaskHistory:
      type: object
      properties:
        id:
          type: integer
          format: int64
        task_id:
          type: integer
          format: int64
        event:
          type: string
        actor:
          type: string
        at:
          type:
            - string
            - "null"
      required: [id, task_id, event, actor]
    TaskDetails:
      type: object
      properties:
        task:
          $ref: "#/components/schemas/Task"
        history:
          type: array
          items:
            $ref: "#/components/schemas/TaskHistory"
      required: [task, history]
    TaskListResponse:
      type: object
      properties:
        ok:
          type: boolean
          const: true
        data:
          type: array
          items:
            $ref: "#/components/schemas/Task"
      required: [ok, data]
    TaskDetailsResponseWrapper:
      type: object
      properties:
        ok:
          type: boolean
          const: true
        data:
          $ref: "#/components/schemas/TaskDetails"
      required: [ok, data]
    CreateTaskRequest:
      type: object
      properties:
        from:
          type: string
        to:
          type: string
        description:
          type: string
      required: [from, to, description]
    AckTaskRequest:
      type: object
      properties:
        id:
          type: integer
          format: int64
        worker:
          type: string
      required: [id, worker]
    CompleteTaskRequest:
      type: object
      properties:
        id:
          type: integer
          format: int64
        worker:
          type: string
        result:
          type: string
      required: [id, worker, result]
    RequeueTaskRequest:
      type: object
      properties:
        id:
          type: integer
          format: int64
      required: [id]
    TaskStatusResponse:
      type: object
      properties:
        ok:
          type: boolean
        id:
          type: integer
          format: int64
      required: [ok, id]
    Message:
      type: object
      properties:
        id:
          type: integer
          format: int64
        sender:
          type: string
        recipient:
          type: string
        body:
          type: string
        created_at:
          type:
            - string
            - "null"
        read:
          type: integer
      required: [id, sender, recipient, body, read]
    MessageListResponse:
      type: object
      properties:
        ok:
          type: boolean
          const: true
        data:
          type: array
          items:
            $ref: "#/components/schemas/Message"
      required: [ok, data]
    SendMessageRequest:
      type: object
      properties:
        from:
          type: string
        to:
          type: string
        body:
          type: string
      required: [from, to, body]
    MarkReadRequest:
      type: object
      properties:
        id:
          type: integer
          format: int64
      required: [id]
    MessageSendResponse:
      type: object
      properties:
        ok:
          type: boolean
        from:
          type: string
        to:
          type: string
      required: [ok, from, to]
    MessageReadResponse:
      type: object
      properties:
        ok:
          type: boolean
        id:
          type: integer
          format: int64
      required: [ok, id]
    GhostOsInferRequest:
      type: object
      properties:
        prompt:
          type: string
      required: [prompt]
    GhostOsInferResponse:
      type: object
      properties:
        output:
          type: string
      required: [output]